requirements.txt
fastapi==0.121.1
uvicorn[standard]==0.38.0
langgraph==1.0.0
langchain[openai]==1.0.0
psycopg[binary,pool]
langgraph-checkpoint-postgres
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.9
email-validator==2.1.1
minio==7.2.5
python-magic==0.4.27

# RAG dependencies
celery[redis]
redis
langchain-core
langchain-postgres
langchain-openai
pypdf
python-docx
tiktoken
pgvector
langchain-community
markdown